32-bit compatibility fix for fix for #5165

+++ test/core.jl
@@ -1185,7 +1185,7 @@ f5150(T) = Array(Rational{T},1)
 # issue #5165
 bitstype 64 T5165{S}
 make_t(x::Int64) = Base.box(T5165{Nothing}, Base.unbox(Int64, x))
-xs5165 = T5165[make_t(1)]
+xs5165 = T5165[make_t(int64(1))]
 b5165 = IOBuffer()
 for x in xs5165
     println(b5165, x)   # segfaulted
